Allow use of arbitrary indent styles
Recognizes different styles of indents e.g. a tab, 3 spaces, 207 spaces
with 15 tabs, etc. The first occurrence of whitespace at the start of the line
will trigger the code to define that collection of whitespace as an "indent."
This can break if a blank line has whitespace characters.

Also fixes the error where "indents" embedded inside the section title
would count towards the depth.